Fix Datetime picker migration
See issue for additional detail #523. Clicking the 'Update Plot' button would cause the selected datetime values to shift by 1 day. This was caused by a UTC issue in which the date was converted to UTC equivalent but then the year, month, and day are later used to repopulate this picker values.