SQLAlchemy ODM2 data models
This implements an SQLAlchemy base approach to ODM2 data models. This approach use the automap_base feature map the database schema to a set of classes in the application. The results of this mapping are also cached in a pickle file so that subsequent builds of the application will read the mappings metadata and generate the data models using a declarative_base approach.