TSV JS Class to manage timeseries
This commit adds a new class which encapsulates the timeseries data into class instances with methods to retrieve more data when necessary. This will allow us to have more a performant application because we can better cache and load only the data that is requested, instead of the full timeseries.