Revised approach to sync dataloader_sensormeasurement table
Related to #560
This revises how the API data stream keeps the dataloader table in sync. The logic has been simplified to only create new records in the dataloader_sensormeasurement table when one does not exist. Separately a database function have been created to keep the sensormeasure table in sync and in triggered whenever data is inserted into the timeseriesresultvalues table.