Hotfix for Deleting sites with data
Related to #574
Due to compression of the timeseriesresultvalues datatable, the command for deleting a site and cascading through various table fails when it arrives at the timeseriesresultvalues table, preventing the site from being deleting. This hot fix solution modifies the constraint so that on delete for timeseriesresultvalues will not cascade. This will result in orphaned timeseriesresultvalues in the database, so this solution is not ideal. Additional work is needed in the near future to develop methods for decompressing the tables and deleting the timeseriesresultvalues in cascade.