avoid accessing ORM during import
A few places make Django ORM queries at the module level and thus run
during import of the portal code. This implies that a working connection
to a properly set-up database is required to e.g. use the Django management
command. This greatly complicates packaging, initial setup, and deployment.

This problem is fixed by simple refactorings and caches.