turn sql model cache creation failure from error to warning
Creation of the SQLAlchemy model caches occurs during import of the
portal code. This implies that a working connection to a properly set-up
database is required to e.g. use the Django management command. This
greatly complicates packaging, initial setup, and deployment.

This problem is fixed by ignoring errors creating the cache and reminding
the user to check the database.