Implementation of has_streamwatch method for SiteRegistration
Reference #645
This commit add a new data field (streamwatch_assessments) to the SiteRegistration object and database table. This field is an integer which tracks the number of StreamWatch assessments completed for a site. Additional database (sql) function were developed and deployed to our database to auto increment and decrement this field when a StreamWatch action is added or removed.