Prep for declaratively mapped models
Previously a new ORM was implemented using SQLAlchemy's automap feature to declare a set of data models. The longer term vision is to do anyway with automapping in favor of a declarative approach which defines the models more explicitly. This commit cleans up the ODM2 namespace but shifting the automap_models to a new portion of the namespace and will allow us to implement new models which declaratively map the new fields.