Filter out accounts with no provided name
#671 - Item 1
The dropdowns are pulling from user account records which are messy. There are accounts with no names provided, inconsistent capitalizations, etc. This is a first pass to at least prevent some of those accounts from appearing in the dropdown. I think the ultimate solution is to clean up the user table, which will come as part of later efforts.