Fix browse sites page
By deprecating some of the fields, we needed to update the database access locations for the browse sites page. The django models are not performant and will be replaced long term, so this just implements a raw sql query to fetch the data in the interim. Needs to be updated to SQLAlchemy ORM once implemented.