Add configurable google maps API key.
The API key was hard coded in browse sites page. This combined with security restrictions that prevent that key from being used outside of specifics domains, made development and testing of map changes difficult. This adds google map key to the configuration to allow local override. This also should allow us to remove hardcoded API from the rest of the code base.