FIx browse sites search feature
Related issue: #689
The search feature still contained a reference to deployed by in the textSearchFacets. That field is removed with site ownership being restructured to an organization, and the field no longer appears in the meta data for the sites used in the map. This was throwing a key error causing the feature to fail.