Restore CSV Download
Related Issue #713
The CSV download view was still pulling affiliation information based on a site registrations account. With the change to organizational ownership, a SiteRegistration's affiliation is not tied to a single account but multiple accounts, and the relationship is no longer 1-to-1 (i.e. site's are affiliated with multiple accounts). For those reasons I removed the affiliation related portions of the metadata causing the error but to a bad (stale) query strucutre.