Change local time labels in CSV
The label was updated to include standard local time, but as @SRGDamia1 noted, the ingestion endpoint is only aware of what the UTC offset is and not what that offset means in the context of local time. Indeed some stations actually switch between local and daylight time. Correcting the label to just local time.